requirements.txt
imageio==2.31.1
imageio-ffmpeg==0.4.8
Pillow==9.5.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
streamlit==1.25.0
streamlit-modal==0.1.0
streamlit-option-menu==0.3.6
streamlit-server-state==0.16.0